fix: speed up loading of authenticated pages
Every feature now keeps its query key factory in a dependency-free
`keys.ts`, and `queries.ts` imports it from there rather than declaring
it. The SSE handler needs keys, not hooks, so it now imports twelve
`keys.ts` modules instead of twelve `queries.ts` modules.

Importing `queries.ts` for a key dragged the whole request layer along
with it. Five feature query modules and four server-function stubs were
sitting in the `_authenticated` chunk purely because the handler
referenced their keys, and through the server-function stubs a second,
duplicate copy of zod. None of it was reachable from anything else in
that chunk, which every authenticated page loads before its own code.

Measured on the production build, the `_authenticated` chunk drops from
97.0 KB to 56.8 KB gzipped. The eager bundle is unchanged: contrary to
what VIR-2679 assumed, the router's `autoCodeSplitting` already keeps
the SSE handler out of it, so `/login` never downloaded any of this. The
eager superagent and zod come from route `loader`/`beforeLoad` halves
importing `queries.ts`, which is a separate problem.

`queries.ts` deliberately does not re-export its keys — a convenience
re-export would be a second, heavier door onto the same value, and the
one an import completion would offer. `reactQueryHandler.test.ts` now
asserts the handler imports no `queries` module, because nothing else
would fail if it did.
