+++ apps/web/src/administration/keys.ts
@@ -1,0 +1,7 @@
+import { createQueryKeys } from "@app/queryKeys";
+
+/** Query keys for the server settings. */
+export const settingsQueryKeys = createQueryKeys("settings");
+
+/** Query keys for administrator roles. */
+export const roleQueryKeys = createQueryKeys("roles");
