+++ apps/web/src/analyses/keys.ts
@@ -1,0 +1,4 @@
+import { createQueryKeys } from "@app/queryKeys";
+
+/** Query keys for analyses. */
+export const analysesQueryKeys = createQueryKeys("analyses");
