+++ apps/web/src/groups/keys.ts
@@ -1,0 +1,4 @@
+import { createQueryKeys } from "@app/queryKeys";
+
+/** Query keys for groups. */
+export const groupQueryKeys = createQueryKeys("groups");
