+++ apps/web/src/hmm/keys.ts
@@ -1,0 +1,4 @@
+import { createQueryKeys } from "@app/queryKeys";
+
+/** Query keys for HMMs. */
+export const hmmQueryKeys = createQueryKeys("hmms");
