+++ apps/web/src/indexes/queries.ts
@@ -1,5 +1,5 @@
 import { apiClient } from "@app/api";
-import { createQueryKeys } from "@app/queryKeys";
+import { indexQueryKeys } from "@indexes/keys";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import type { ErrorResponse } from "@/types/api";
 import type {
@@ -8,23 +8,6 @@ import type {
 	IndexSearchResult,
 	UnbuiltChangesSearchResults,
 } from "./types";
-
-/**
- * Factory for generating react-query keys for index related queries.
- */
-const indexKeys = createQueryKeys("indexes");
-
-/**
- * Query keys for indexes.
- *
- * `unbuilt()` is the list of a reference's changes that no index covers yet. It
- * is keyed by reference rather than by index, so it gets its own namespace
- * instead of sharing the index details one.
- */
-export const indexQueryKeys = {
-	...indexKeys,
-	unbuilt: (refId: string) => [...indexKeys.all(), "unbuilt", refId] as const,
-};
 
 /**
  * Gets a paginated list of indexes.
