+++ apps/web/src/jobs/keys.ts
@@ -1,0 +1,4 @@
+import { createQueryKeys } from "@app/queryKeys";
+
+/** Query keys for jobs. */
+export const jobQueryKeys = createQueryKeys("jobs");
