+++ apps/web/src/labels/keys.ts
@@ -1,0 +1,4 @@
+import { createQueryKeys } from "@app/queryKeys";
+
+/** Query keys for labels. */
+export const labelQueryKeys = createQueryKeys("labels");
