+++ apps/web/src/references/keys.ts
@@ -1,0 +1,4 @@
+import { createQueryKeys } from "@app/queryKeys";
+
+/** Query keys for references. */
+export const referenceQueryKeys = createQueryKeys("references");
