+++ apps/web/src/samples/keys.ts
@@ -1,0 +1,4 @@
+import { createQueryKeys } from "@app/queryKeys";
+
+/** Query keys for samples. */
+export const samplesQueryKeys = createQueryKeys("samples");
