+++ apps/web/src/subtraction/queries.ts
@@ -1,5 +1,5 @@
 import { apiClient } from "@app/api";
-import { createQueryKeys } from "@app/queryKeys";
+import { subtractionQueryKeys } from "@subtraction/keys";
 import {
 	keepPreviousData,
 	useMutation,
@@ -12,20 +12,6 @@ import type {
 	SubtractionOption,
 	SubtractionSearchResult,
 } from "./types";
-
-const subtractionKeys = createQueryKeys("subtractions");
-
-/**
- * Query keys for subtractions.
- *
- * `shortlist()` is the reduced list used to populate selectors. It nests under
- * `lists()` so that a list invalidation refreshes it too.
- */
-export const subtractionQueryKeys = {
-	...subtractionKeys,
-	shortlist: (ready?: boolean) =>
-		[...subtractionKeys.lists(), "short", ready] as const,
-};
 
 /**
  * Initializes a mutator for creating a subtraction
