+++ apps/web/src/tasks/keys.ts
@@ -1,0 +1,4 @@
+import { createQueryKeys } from "@app/queryKeys";
+
+/** Query keys for tasks. */
+export const taskQueryKeys = createQueryKeys("tasks");
