+++ apps/web/src/wall/keys.ts
@@ -1,0 +1,4 @@
+import { createQueryKeys } from "@app/queryKeys";
+
+/** Query keys for the root document. */
+export const rootQueryKeys = createQueryKeys("root");
